UBUNTU: [Debian] Work out upstream tag for use with gen-auto-reconstruct
The current method of invoking gen-auto-reconstruct doesn't work
when the development kernel is based off of a -rc upstream
version. Not that autoreconstruct is essential when we don't have
an orig tarball, however insertchanges invokes autoreconstruct so
it's inconvenient for it to fail.

Update 0-common-vars.mk to generate a variable representing the
upstream tag for the base version for this kernel. Then change
the autoreconstruct target to pass this tag to
get-auto-reconstruct so that it will work even for kernels
based on release candidates.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>