UBUNTU: vbox-update: Fix up KERN_DIR definitions
The 5.2.0-dfsg-2 vbox package has some makefile changes which
define KERN_DIR under the asumption that modules are being built
against external headers found under /lib/modules. This is not
true when building the modules alongside the kernel, and the vbox
build aboarts because the path doesn't exist.

Update vbox-update to automatically replace the KERN_DIR
definitions with one which points it at the base of the kernel
source tree.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>