UBUNTU: [Debian] Fix invocation of dh_prep for dbgsym packages
Currently when ivoking dh_prep for dbgsym packages we use a
nonexistant $dbg_pkg_name variable. This has been the case for
a long time, but it never caused problems until debhelper was
updated in bionic.

We can't simply define $dbg_pkg_name because 'dbgsym' in the
package names comes after the flavor. Fix up the rule to use
the correct name for the dbgsym package, based on the binary
package name.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>