UBUNTU: SAUCE: (efi-lockdown) Lock down TIOCSSERIAL
Lock down TIOCSSERIAL as that can be used to change the ioport and irq
settings on a serial port.  This only appears to be an issue for the serial
drivers that use the core serial code.  All other drivers seem to either
ignore attempts to change port/irq or give an error.

Reported-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: David Howells <dhowells@redhat.com>
cc: Jiri Slaby <jslaby@suse.com>
(cherry picked from commit aa1bba375c991422b971fd90a7bf602046e5f1a8
 git://git.kernel.org/pub/scm/linux/kernel/git/jforbes/linux.git)
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>