UBUNTU: change i386 build_arch to i386 instead of x86
Upstream commit f467c5640c29ad258c3cd8186a776c82fc3b8057 ("kconfig: only write
'# CONFIG_FOO is not set' for visible symbols") causes CONFIG_64BIT to not be
set when ARCH=i386 because then it defaults to n.

config 64BIT
        bool "64-bit kernel" if ARCH = "x86"
        default ARCH != "i386"

As we generate the configs by using ARCH=i386, when we run updateconfigs, we
end up with CONFIG_64BIT not even appearing in the output file. This should not
be a problem as it's going to be calculated as 'n' when we run silentoldconfig
again.

However, as we use ARCH=x86 when we build the kernel, we end up with a
different config. If we set build_arch to i386 on the arch hook, that is fixed,
and we don't have that problem anymore.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>