UBUNTU: [Config] Fix SOF Kconfig options
BugLink: https://bugs.launchpad.net/bugs/1848490

As reported on the above bug and elsewhere, some incorrect config
options have resulted in non-working sound for some laptops. In
eoan we're applying a more conservative set of config changes,
but for focal we'll apply the set of changes which was
recommended to us by one of the upstream Intel ASoC maintainers.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>