UBUNTU: [Config] FRAMEBUFFER_CONSOLE_DEFERRED_TAKEOVER=y
BugLink: https://bugs.launchpad.net/bugs/1836858

Impact:

Flickerfree boot doesn't work with Eoan and Focal kernels - the flicker
occurs seemingly during kernel startup following this pattern:

1. BIOS logo on a black background
2. Display mode resets, no image on a black background briefly
3. BIOS logo on a black background

According to upstream, the required kernel option to achieve it (coupled
with the quiet cmdline) is FRAMEBUFFER_CONSOLE_DEFERRED_TAKEOVER.

Fix:

Apply the config change, recompile and reboot - a test kernel was
provided and the reporter of this bug said "it doesn't completely
solve the issue, but it's a step in the right direction, so please
enable it".

Regression potential:

Low, config change for an option that is upstream.

Proposing for Focal.

Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>
[ saf: update annotations ]
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>