UBUNTU: SAUCE: libtraceevent: Strip symbol version from nm output
BugLink: https://bugs.launchpad.net/bugs/1887397

The following change in binutils causes 'nm -D' output to include
symbol version:

https://sourceware.org/git/gitweb.cgi?p=binutils-gdb.git;h=1081065c519d1bfc3847bf4b0a0ce4bc3224bcd3

This causes libtraceevent-dynamic-list to also include the
versions, which results in a link error. Get around this by
stripping off any symbol version information, if present, which
follows the symbol name and is separated from the symbol name by
one or two @ characters.

Note that I am not certain whether or not the use of gsub() for
stripping the symbol version is compatible with all awk
implementations.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>