UBUNTU: [Packaging] Bind hv_vss_daemon startup to hv_vss device
BugLink: https://bugs.launchpad.net/bugs/1891224

Update the systemd service for the vss daemon and add a new udev rule
to ensure the service will not fail in case vss is not enabled for the
guest.

The udev rules is responsible for starting the service only when the
vss device is available. In the unit file, the service is now bound to
the device unit, so the service can be stopped if the device is
removed. And the service is also conditioned to the existence of the
device, this way we avoid starting and stopping the service during
boot if the device is not available.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>