UBUNTU: [Config] Update numerous configs to conform with policy
When reviewing the annotations updates for the 5.10-rc2 rebase,
I noted a large number of options which did not conform to our
config policy. These have been updated. I suspect there may be
others from the 5.10-rc1 rebase which also do not conform to
policy, so further review is needed.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>