UBUNTU: [Config] Deactivate CONFIG_QETH_OSN kernel config option
BugLink: https://bugs.launchpad.net/bugs/1906370

The final shipment of OSN devices was as part of the IBM z13 hardware generation.
The primary exploiter was the IBM Communication Controller,
which was pulled out of marketing in March 2015 and should be out of service now.
Therefore, IBM pulls the support from all Linux distros going forward.
Hence the deactivation of the CONFIG_QETH_OSN kernel config option for hirsute and onwards.

Signed-off-by: Frank Heimes <frank.heimes@canonical.com>
[ saf: fix ups ]
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>