UBUNTU: SAUCE: Revert "UBUNTU: SAUCE: overlayfs: Be more careful about copying up sxid files"
Copy up now happens under the creator creds (or under creds provided
by an LSM), meaning that it is no longer possible that a sxid file
could be copied up if the mounter could not have created the file.
Therefore our sauce patch is no longer needed.

Additionally, the operations used by the exploits for CVE-2016-1575
and CVE-2016-1576 now fail in permission checks prior to any attempt
to copy up from the lower filesystem, which also prevents the attacks.

I have confirmed that the original exploits no longer work with this
patch reverted.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>