UBUNTU: SAUCE: Revert "UBUNTU: SAUCE: overlayfs: Propogate nosuid from lower and upper mounts"
Without this patch it will be posible to mount an overlay filesystem
without nosuid over a filesystem with nosuid. Generally this is not
an issue, since it will only be possible to setuid to ids within the
user namespace of the mounter, over which the mounter already has
privileges.

The path to the exploit was by triggering copy-up of the file from the
lower filesystem into an upper directory in a mount without nosuid.
This will now fail permission checks prior to copy up, and even if it
did not the copy up would happen under the creator creds, which would
not permit creating a setuid file to a user over which the mounter
does not already have privileges.

I have tested the original exploit with this patch reverted to confirm
that the problematic copy up is not permitted.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>