UBUNTU: SAUCE: ceph: dump info about cap flushes when we're waiting too long for them
BugLink: https://bugs.launchpad.net/bugs/2032959

We've had some cases of hung umounts in teuthology testing. It looks
like client is waiting for cap flushes to complete, but they aren't.

Add a field to the inode to track the highest cap flush tid seen for
that inode. Also, add a backpointer to the inode to the ceph_cap_flush
struct.

Change wait_caps_flush to wait 60s, and then dump info about the
condition of the list.

Also, print pr_info messages if we end up dropping a FLUSH_ACK for an
inode onto the floor, or if we get a message on an unregistered
session.

Reported-by: Patrick Donnelly <pdonnell@redhat.com>
URL: https://tracker.ceph.com/issues/51279
Signed-off-by: Jeff Layton <jlayton@kernel.org>
(cherry picked from https://github.com/ceph/ceph-client.git)
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>