UBUNTU: SAUCE: Revert "usb: core: reduce power-on-good delay time of root hub"
BugLink: https://bugs.launchpad.net/bugs/1939638

This reverts commit 49b64d3f82f33f9bdd313538795dc00c3f85090c.

The USB devices will fail to be detected if the power-on-good delay
of the root hub is not long enough. In the problematic root hub,
it claims 20 ms of bPwrOn2PwrGood in the hub descriptor, but we
have to increase it to 100ms to make all connected USB devices to
be correctly enumerated and detected. Revert the commit until a
formal fix from upstream.

Signed-off-by: Chris Chiu <chris.chiu@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>