UBUNTU: [Packaging] switch to kernel-versions
Switch to obtaining the DKMS package versions from the kernel-versions
dataset rather than from the archive.  This allows it to be more
resilient against parallel update of those versions in the archive.

Replace the existing `update-versions-dkms` script with
`update-dkms-versions`.  This change in name is deliberate as the new
script must be called at a different stage of the crank process, it must
follow the `cranky link-tb` stage to obtain the correct versions.  See
the crank documentation for details.

BugLink: https://bugs.launchpad.net/bugs/1928921
Properties: no-test-build
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>