UBUNTU: [Packaging] workaround missing python3-dev dependency
The following commit is meant to fix build issues in environment where
python2 runtime is only installed:

 630af16eee49 ("perf tools: Use Python devtools for version autodetection rather than runtime")

However, this requires python3-dev that is not specified in the build
dependencies.

As a temporary workaround set PYTHON=python3 in debian/rules to enforce
the usage of python3 when python is needed.

TODO: add python3-dev to the build dependencies and update all our build
chroots.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>