UBUNTU: [Config] Set CONFIG_PWM_SIFIVE=m for riscv64
BugLink: https://launchpad.net/bugs/1988984

Only the SoCs on the SiFive Unleashed and Unmatched boards use this
driver, but none of them needs it at boot time. So compile it as a
module to save memory on other systems.

This also works around a bug in the UART driver fixed by
643792048ee8 ("serial: sifive: enable clocks for UART when probed")

Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>